可视化-各来源流量统计-1

+++ online-education-publisher/src/main/java/com/atguigu/online/education/bean/TrafficDurPerSession.java
@@ -3,6 +3,8 @@ package com.atguigu.online.education.bean;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 // 流量域 各会话页面访问时长
 @Data
 @AllArgsConstructor
@@ -10,5 +12,5 @@ public class TrafficDurPerSession {
     // 来源
     String sc;
     // 各会话页面访问时长
-    Double durPerSession;
+    BigDecimal durPerSession;
 }
