+++ online-education-publisher/src/main/java/com/atguigu/online/education/bean/TrafficPvPerSession.java
@@ -3,6 +3,8 @@ package com.atguigu.online.education.bean;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 // 流量域 各会话页面浏览数
 @Data
 @AllArgsConstructor
@@ -10,5 +12,5 @@ public class TrafficPvPerSession {
     // 来源
     String sc;
     // 各会话页面浏览数
-    Double pvPerSession;
+    BigDecimal pvPerSession;
 }
